+++ scrapyd/website/__init__.py
@@ -1,0 +1,1 @@
+from root import Root